# notification-service/src/main/java/com/loadup/notification/dto/Customer.java
package com.loadup.notification.dto;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class Customer {

    private String customerId;

    private CommunicationPreference communicationPreference;

    private String address;

    private String email;

    @Getter
    public enum CommunicationPreference {

        EMAIL("EMAIL"),

        MAIL("MAIL");

        private String value;

        CommunicationPreference(String value) {
            this.value = value;
        }

    }
}