# order-service-app/src/main/java/com/loadup/order/repository/model/OrderStatus.java
package com.loadup.order.repository.model;

public enum OrderStatus {
    CREATED,
    CONFIRMED,
    PAID,
    FULFILLED,
    SHIPPED,
    DELIVERED,
    CANCELLED,
    UPDATED,
    COMPLETED,
    FAILED;
}